+++ src/components/Sidebar.js
@@ -6,7 +6,7 @@ import "../styles/Sidebar.css"; // Dedicated CSS for the layout and side panel
 import { useStreamChat } from "../context/StreamChatContext";
 // Import FontAwesome icons
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faUser, faHome, faCog, faComment, faRobot, faSignOutAlt, faBell, faUsers } from "@fortawesome/free-solid-svg-icons";
+import { faUser, faHome, faCog, faComment, faRobot, faSignOutAlt, faBell, faUsers, faStar } from "@fortawesome/free-solid-svg-icons";
 
 const Sidebar = () => {
   const navigate = useNavigate();
@@ -63,11 +63,18 @@ const Sidebar = () => {
               </NavLink>
             </li>
             <li>
+                                      <NavLink to="/points" className="sidebar-menu-link" activeClassName="sidebar-active-link">
+                                        <FontAwesomeIcon icon={faStar} />
+                                        Points
+                                      </NavLink>
+                                    </li>
+            <li>
               <NavLink to="/settings" className="sidebar-menu-link" activeClassName="sidebar-active-link">
                 <FontAwesomeIcon icon={faCog} />
                 Settings
               </NavLink>
             </li>
+
           </ul>
 
           {/* Logout Button */}